+++ cli/agent.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"github.com/drewsonne/gocdsdk"
+	"github.com/drewsonne/go-gocd/gocd"
 	"github.com/urfave/cli"
 )
 
@@ -113,17 +113,19 @@ func DeleteAgentsAction(c *cli.Context) error {
 
 func ListAgentsCommand() *cli.Command {
 	return &cli.Command{
-		Name:   ListAgentsCommandName,
-		Usage:  ListAgentsCommandUsage,
-		Action: ListAgentsAction,
+		Name:     ListAgentsCommandName,
+		Usage:    ListAgentsCommandUsage,
+		Action:   ListAgentsAction,
+		Category: "Agents",
 	}
 }
 
 func GetAgentCommand() *cli.Command {
 	return &cli.Command{
-		Name:   GetAgentCommandName,
-		Usage:  GetAgentCommandUsage,
-		Action: GetAgentAction,
+		Name:     GetAgentCommandName,
+		Usage:    GetAgentCommandUsage,
+		Action:   GetAgentAction,
+		Category: "Agents",
 		Flags: []cli.Flag{
 			cli.StringFlag{Name: "uuid, u", Usage: "GoCD Agent UUID"},
 		},
@@ -132,9 +134,10 @@ func GetAgentCommand() *cli.Command {
 
 func UpdateAgentCommand() *cli.Command {
 	return &cli.Command{
-		Name:   UpdateAgentCommandName,
-		Usage:  UpdateAgentCommandUsage,
-		Action: UpdateAgentAction,
+		Name:     UpdateAgentCommandName,
+		Usage:    UpdateAgentCommandUsage,
+		Action:   UpdateAgentAction,
+		Category: "Agents",
 		Flags: []cli.Flag{
 			cli.StringFlag{Name: "uuid, u", Usage: "GoCD Agent UUID"},
 			cli.StringFlag{Name: "config, c", Usage: "JSON encoded config for agent update."},
@@ -144,9 +147,10 @@ func UpdateAgentCommand() *cli.Command {
 
 func DeleteAgentCommand() *cli.Command {
 	return &cli.Command{
-		Name:   DeleteAgentCommandName,
-		Usage:  DeleteAgentCommandUsage,
-		Action: DeleteAgentAction,
+		Name:     DeleteAgentCommandName,
+		Usage:    DeleteAgentCommandUsage,
+		Action:   DeleteAgentAction,
+		Category: "Agents",
 		Flags: []cli.Flag{
 			cli.StringFlag{Name: "uuid, u", Usage: "GoCD Agent UUID"},
 		},
@@ -155,9 +159,10 @@ func DeleteAgentCommand() *cli.Command {
 
 func UpdateAgentsCommand() *cli.Command {
 	return &cli.Command{
-		Name:   UpdateAgentsCommandName,
-		Usage:  UpdateAgentsCommandUsage,
-		Action: UpdateAgentsAction,
+		Name:     UpdateAgentsCommandName,
+		Usage:    UpdateAgentsCommandUsage,
+		Action:   UpdateAgentsAction,
+		Category: "Agents",
 		Flags: []cli.Flag{
 			cli.StringSliceFlag{Name: "uuid", Usage: "GoCD Agent UUIDs"},
 			cli.StringFlag{Name: "state", Usage: "Whether agents are enabled or disabled. Allowed values 'Enabled','Disabled'."},
@@ -168,9 +173,10 @@ func UpdateAgentsCommand() *cli.Command {
 
 func DeleteAgentsCommand() *cli.Command {
 	return &cli.Command{
-		Name:   DeleteAgentsCommandName,
-		Usage:  DeleteAgentsCommandUsage,
-		Action: DeleteAgentsAction,
+		Name:     DeleteAgentsCommandName,
+		Usage:    DeleteAgentsCommandUsage,
+		Action:   DeleteAgentsAction,
+		Category: "Agents",
 		Flags: []cli.Flag{
 			cli.StringSliceFlag{Name: "uuid", Usage: "GoCD Agent UUIDs"},
 		},