+++ pipeline.go
@@ -1,5 +1,0 @@
-package gocd
-
-type Pipeline struct {
-	Name string `json:"name"`
-}